Organised front end layout
