+++ frontend/src/assets/styles/main.css
