+++ frontend/src/components/chat/CoachChatInterface.vue
