+++ frontend/src/components/chat/RefinerChatInterface.vue
