+++ frontend/src/views/CoachView.vue
@@ -1,0 +1,7 @@
+<template>
+  <CoachChatInterface />
+</template>
+
+<script setup>
+import CoachChatInterface from '@/components/chat/CoachChatInterface.vue'
+</script>
