--- frontend/src/views/CoachingView.vue
@@ -1,7 +1,0 @@
-<template>
-  <CoachChatInterface />
-</template>
-
-<script setup>
-import CoachChatInterface from '@/components/CoachChatInterface.vue'
-</script>
